refactor: (#56) features의 game 슬라이스의 ui import 순서 및 경로 변경 (순환 참조 방지)
