+++ src/entities/game/index.ts
@@ -1,3 +1,5 @@
 export * from './model/types';
+// TODO: ItemIcon 데이터 전용 타입 생성 예정.
+export type { ItemIconProps } from './ui/ItemIcon';
 export { ItemIcon } from './ui/ItemIcon';
 export { PlayerAvatar } from './ui/PlayerAvatar';
