+++ src/features/game/ui/InventoryPanel.tsx
@@ -1,9 +1,10 @@
-import type { GameItem } from '@/entities/game/model';
-import { useInventory } from './useInventory';
-import { ItemIcon, type ItemIconProps } from '@/entities/game/ui/ItemIcon';
-import { COLORS, UI_TEXT } from '@/features/game/constants/game';
-import { ChevronUpIcon, ChevronDownIcon } from '@heroicons/react/24/solid';
+import { ChevronDownIcon, ChevronUpIcon } from '@heroicons/react/24/solid';
+
+import type { GameItem, ItemIconProps } from '@/entities/game';
+import { ItemIcon } from '@/entities/game';
 import { cn } from '@/shared/lib/cn';
+import { COLORS, UI_TEXT } from '../constants/game';
+import { useInventory } from './useInventory';
 
 interface InventoryPanelProps {
   inventory: GameItem[] | null;
@@ -37,6 +38,7 @@ export const InventoryPanel = ({
         </button>
 
         <div className="flex flex-col gap-6 my-2 h-[450px]">
+          {/* TODO: 데이터 전용 타입 설계 */}
           {visibleItems.map((item: ItemIconProps) => (
             <ItemIcon key={item.id} {...item} />
           ))}
